Move to PostgreSQL and drop SQLite
SQLite was always the temporary engine. Last round isolated every
dialect-dependent decision, so this one landed where it was supposed to:
dialect.go and migrations/ were rewritten, and store.go's 1500 lines were
not touched at all.

- dialect.go now speaks pgx and SQLSTATE. isForeignKeyViolation and
  isRestrictViolation collapse onto 23503 — PostgreSQL does not
  distinguish them, so the direction of the call is what carries the
  meaning, exactly as the two functions were already used.
- Migrations rewritten: BIGINT timestamps (Unix seconds overflow int4 in
  2038), BOOLEAN where Go writes a bool, IDENTITY for the audit log's
  key, and lower() expression indexes where COLLATE NOCASE used to be.
  The rationale comments were kept; only the dialect changed.
- store.Open takes a connection string and configures a pool. Absent an
  explicit sslmode it forces verify-full: libpq defaults to "prefer",
  which silently falls back to plaintext, and a bastion's identity and
  audit traffic is not the place to accept a downgrade.
- database.path became database.dsn, overridable by POSTERN_DATABASE_DSN
  so the password need not sit in a config file. Unlike every other
  secret it cannot live in the settings table: it is what gets us there.

Migrate now holds a PostgreSQL advisory lock. The hazard is new — a
single file gave one writer for free, a shared server does not. Verified
by removing the lock: three of four concurrent runners fail with a
duplicate key on pg_type_typname_nsp_index, a catalog-level collision
that tells an operator nothing and leaves a half-built schema.

Tests run against a real PostgreSQL (internal/testdb): one container per
test binary, a schema per test. Testing the store against anything else
would test the wrong thing. Without Docker they fail rather than skip, so
a green run means the store was actually exercised; -short opts out.

+++ internal/store/migrations/005_role_source.up.sql
@@ -10,12 +10,11 @@
 -- Süresi dolan satır SİLİNMEZ: denetim izi olarak kalır ("bu yetki vardı,
 -- doldu") ve sorgular tembel filtreyle görmezden gelir.
 
--- SQLite ALTER TABLE tek seferde tek sütun ekler.
 ALTER TABLE user_roles ADD COLUMN source TEXT NOT NULL DEFAULT 'manual'
   CHECK (source IN ('sso', 'manual'));
 
 -- NULL = süresiz. Unix saniyesi (şemanın geri kalanıyla aynı biçim).
-ALTER TABLE user_roles ADD COLUMN expires_at INTEGER;
+ALTER TABLE user_roles ADD COLUMN expires_at BIGINT;
 
 -- "Bu kullanıcının SSO rollerini yenile" en sık çalışacak sorgu.
 CREATE INDEX user_roles_source_idx ON user_roles(user_id, source);
